// src/app/pages/table-cliet/client.interface.ts
export interface Client {
  firstName?: string;
  lastName?: string;
  email: string;
  phone: string;
  address: string;
  region: string;
  codePostal: string;
  companyName: string;
  localOrshore?: string;
  tva?: string;
  etat?: string;
  fax?: string;
  ibanRib?: string;
  swiftBic?: string;
  nattestation?: string;
  codeFiscal?: string;
  conPayment?: string;
  website?: string;
  techContact?: string;
  nRegisterCommerce?: string;
}